SLProgram: move return info out of lines, into .ret field
This brings some simplifications, and will allow two SLProgram
to share the same lines but return different values.
The alternative is to store the return info together with
the upcoming `.ptr` field packed in the final `Line`,
but then this has special status and the invariant that
this must appear only at the end must be checked/enforced.
Better to store separately.