Free: compose: add "flatten-mode" to avoid getting a `Compose` object
I.e. we use evalutate to get the result of compose directly,
instead of using a "lazy" Compose structure.
There is currently no "lazy" mode for SLProgram and it's unclear
whether such a mode is useful, so let's have a non-lazy mode for
`Free` to keep 1-1 correspondence with SLProgram (better for
tests among other things).