Transform form (#307)
The novelty of this branch is the file `src/Groups/matrices/transform_form.jl`. This contains an algorithm to compute the change of basis matrix between two forms. At the end of the file there is the upper-level function `iscongruent`. Given forms with Gram matrices `B1`, `B2`, the function `iscongruent` returns (`true`, `C`) if `C` is a matrix such that `C*B1*conjugate_transpose(C) = B2`. If such `C` does not exists, return (`false`, `nothing`).
It works with the following forms:

- alternating forms
- symmetric forms
- hermitian forms
- quadratic forms

It works also for degenerate forms.
The employed algorithm does not cover quadratic forms, so in even characteristic the function `iscongruent` calls the methods from the GAP package "forms" (this is not necessary for odd characteristic, since quadratic forms can be paired with corresponding symmetric forms).