changed the `rand` methods for groups and cosets ...
... such that a (Julia) random number generator can be given
as the first argument.
(This feature relies on the recent release of GAP.jl 0.6.)

The default rng is `Random.GLOBAL_RNG`.
Note that this is not GAP's default rng,
thus the same group related computations in Oscar and GAP
will yield different results if random elements are involved.

I have also changed `GapObj` and `FFE` to `GAP.GapObj` and `GAP.FFE`
in a few tests; up to now, I got test failures when I ran the Groups
tests without calling `using GAP` in advance.