first attempt for `iso_gap_oscar`
Provide `iso_gap_oscar`:
The idea is to compute a `MapFromFunction` object from a given GAP domain,
with `codomain` value an Oscar object.

Currently GAP's finite fields and cyclotomic fields are supported.
However, finite fields that are extensions of non-prime fields
are currently not supported.
(The straightforward approach would be to compute the minimal polynomial
of a primitive element w.r.t. the prime field, and a corresponding basis.
However, this seems to be not reasonable for GAP's `AlgebraicExtension`s.)

Some more remarks:

- The return value is cached on the GAP side, via a GAP attribute.
  In order to provide this attribute, some GAP code must be executed
  at runtime; I do not know a better way.

- It would be tempting to use GAP's method selection for administrating
  the methods for various kinds of GAP domains.
  This would work if one could be sure that the distinguishing flags
  are always stored in the domains, and if the ranking of the GAP methods
  would be safe.
  Currently there is just one method that calls a Julia function which
  runs through the available cases.

- Perhaps the name `iso_gap_oscar` is not a good choice.
  For the other direction, `ring_iso_oscar_gap` had been changed to
  `iso_oscar_gap` because the Julia types of the argument determines
  which kind of isomorphism is intended; this argument does not hold
  for the direction from GAP to Oscar.

- The code around `ring_iso` objects was originally stored in
  `src/Groups/matrices`.
  However, it does not really belong to (matrix) groups, thus it should
  better go to `src/GAP`, as well as its tests.