create Oscar groups with context
(This addresses issue #894.)

Several types of Oscar groups need context information besides the underlying GAP object:
`PermGroup` needs the degree, `MatrixGroup` needs the dimension and the ring of the matrix entries.
This context was not used in certain situations.

Now the recommended way to construct groups with context is via `oscar_group(gapobj, G)` where `gapobj` is a GAP group and `G` is an Oscar group that provides the context.
(Up to now we had `T(gapobj)`, where `T` is the type of `G`, but this does not work in general. It would be tempting to use `T(gapobj, G)`, but this syntax cannot be used in general because `AutomorphismGroup` uses it for something different.)

(More methods should get restricted to situations where the contexts of the arguments are equal. This will be the next step.)