do not set globals in `setup` code of `jldoctest`
Thus setting/resetting the `is_unicode_allowed` flag is now visible
in the docstring examples.

Still it might be a bad idea to show unicode in the manual,
whereas it is switched off in the session by default.
On the other hand, unicode looks just nicer, in particular in the manual.