Fix `stabilizer` for "large" permutation groups
In some cases, computing a stabilizer in a GAP group, e.g. via
`stabilizer(G,1)`, could result in an error inside GAP. This only
happened when the permutation group already knew its size. That could
lead to confusing situations where at first `stabilizer` works, but
subsequent ones fail (because as a side effect of the first stabilizer
computation, the group order was calculated and stored). This patch
fixes this, and adds some tests to verify the fix works as intended.