Align all_*_groups methods some more (#3433)
For `all_small_groups` we allow specifying the order as a single
integer or abstract vector, followed by any number of additional
filters. For transitive and primitive groups, we also allowed
specifying the degree as a single integer or abstract vector, but
this then could not be followed by filters. With this patch the
additional filters work in all cases.