Remove a `sub` method for `MatrixGroup` that is worse than the generic one (#5373)
* remove a `sub` method for `MatrixGroup`

The generic `sub` method is good enough.
The special method for `MatrixGroup` did not support
the `check` keyword argument.

(Changed behaviour due to a new GAP version uncovered
that something was wrong with a test example;
this example is now better than it was before.)

* change a test

The `gens` field gets computed on demand for a `GAPGroup`,
now also for a `MatrixGroup` constructed with `sub`.

---------

Co-authored-by: ThomasBreuer <sam@math.rwth-aachen.de>